LayoutTests/fast/forms/range/range-default-value.html

<html
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<body>
<p id="description"></p>
<form>
<input id="slider" type="range" min="1" max="101">
</form>
<div id="console"></div>
<script>
description("<a href='http://bugs.webkit.org/show_bug.cgi?id=20134'>Bug 20134: REGRESSION: (2.0.4-3.0.4): No default value set for &lt;input type=range&gt; with an even difference of (max - min)</a>");

function quoteNumber(n)
{
    return "\"" + n + "\"";
}

var slider = document.getElementById("slider");
shouldBe("slider.value", quoteNumber(51));
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
